+++ src/main/java/kr/codeit/relaxtogether/repository/GatheringRepository.java
@@ -1,6 +1,6 @@
 package kr.codeit.relaxtogether.repository;
 
-import kr.codeit.relaxtogether.entity.Gathering;
+import kr.codeit.relaxtogether.entity.gathering.Gathering;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GatheringRepository extends JpaRepository<Gathering, Long> {
