+++ src/main/java/kr/codeit/relaxtogether/repository/UserGatheringRepository.java
@@ -1,6 +1,6 @@
 package kr.codeit.relaxtogether.repository;
 
-import kr.codeit.relaxtogether.entity.UserGathering;
+import kr.codeit.relaxtogether.entity.gathering.UserGathering;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserGatheringRepository extends JpaRepository<UserGathering, Long> {
